Update arma.spec.R
I think that it would be a good practice that since you allow a frequency parameter that you also normalize the variance by the period. Then you need to divide, in line 30, var.noise by xfreq (which I think you renamed the frequency parameter input to), ergo, normalizing everything to the frequency basis. 

Another option is to do this in the documentation and inform that var.noise should be divided by frequency, but it would be nice if you could do it for the user.